+++ components/rootscreen/api/src/main/kotlin/com/flipperdevices/rootscreen/model/RootScreenConfig.kt
@@ -34,4 +34,7 @@ sealed class RootScreenConfig {
 
     @Serializable
     data class Changelog(val updateRequest: UpdateRequest) : RootScreenConfig()
+
+    @Serializable
+    data object RemoteControls : RootScreenConfig()
 }